치명타,회피구현 및 DamageProcess.cs생성
